+++ pkg/platform/resource_global_role.go
@@ -58,7 +58,7 @@ var globalRoleActions []string = []string{
 var _ resource.Resource = (*globalRoleResource)(nil)
 
 type globalRoleResource struct {
-	ProviderData PlatformProviderMetadata
+	ProviderData util.ProviderMetadata
 	TypeName     string
 }
 
@@ -195,7 +195,7 @@ func (r *globalRoleResource) Configure(ctx context.Context, req resource.Configu
 	if req.ProviderData == nil {
 		return
 	}
-	r.ProviderData = req.ProviderData.(PlatformProviderMetadata)
+	r.ProviderData = req.ProviderData.(util.ProviderMetadata)
 }
 
 func (r *globalRoleResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {