+++ sample.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     platform = {
       source  = "jfrog/platform"
-      version = "1.9.0"
+      version = "2.0.0"
     }
   }
 }